:truck: refactor: RenjuGameRule 및 GameRule로 Board 내부 로직 변경
